requirements.txt
fastapi>=0.95,<1.0
uvicorn>=0.22,<1.0
sqlmodel>=0.0.8,<1.0
python-dotenv>=1.0.0,<2.0
alembic>=1.12,<2.0
tiktoken>=0.5,<1.0
langgraph>=1.0.0,<2.0
requests>=2.28,<3.0
bcrypt>=4.0.0,<5.0
numpy>=1.25,<3.0
PyJWT>=2.8,<3.0
# Optional: faiss-cpu may not be available on all platforms; install if desired
faiss-cpu>=1.7.4
sentence-transformers>=5.2.2
pypdf>=4.3.1,<6.0
ebooklib>=0.18,<1.0
beautifulsoup4>=4.12,<5.0
python-multipart>=0.0.9,<1.0